pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbt-autodoc"
version = "1.0.0"
description = "Auto-generate dbt model and column descriptions using heuristics or LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
all = ["anthropic>=0.40.0", "openai>=1.50.0"]
dev = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pytest>=7.0",
    "ruff>=0.4.0",
]

[project.scripts]
dbt-autodoc = "dbt_autodoc.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"
